Refactor test_service to have separate tests on MatProcessor
